requirements.txt
captum==0.7.0
matplotlib==3.10.1
numpy==2.2.4
opencv_python==4.11.0.86
Pillow==11.2.0
scikit_learn==1.6.1
seaborn==0.13.2
timm==1.0.15
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
